Second Look
After re evaluating the instructions I believe that the requirements are to use internationalization to display the label text, in whatever locale, while
simultaneously allowing dynamic changes to the Cronut Widget. I think.

As such, I'll be adding a new French locale, in addition to utilizing i18n to swap out the query strings in the label.