Added the actual IFrame, lost some tests, styling
I switched it up and decided to actually create the Iframe rather than just imagine it. There were a few headaches thanks to my choice to render as a JS format,
requiring headers, cross-origin warnings, etc. etc. But it came out rather nice.

	-Placing the iframe tag in a page allows calls the embed method, which wraps the show method up into an iframe and delivers it.
	-The show method had to be wrapped because it has an ajax call that allows dynamic language choices.
	-The language changing links no longer destroy the cronuts_url and donuts_url as it's remote.
I lost a few tests due to the switch to an IFrame but they were redundant and still passing in the show specs.
Security is a non-issue as there is no posting and no potential forgery... And not to mention IFrames are secure.
If put into production the src url in the Iframe needs an update.
But as far as I can reckon, the widget will embed and display the cronut choices, any cronut or donut choices, from anywhere.

TODO - continue testing the embedding, try a live test for funsies.
TODO - You still need to test for broken links in the url params.